feed 去掉对user的引用

+++ feed/service-feed/src/main/java/cn/momia/service/feed/comment/FeedCommentService.java
@@ -1,10 +1,12 @@
 package cn.momia.service.feed.comment;
 
+import cn.momia.api.feed.dto.FeedComment;
+
 import java.util.List;
 
 public interface FeedCommentService {
     boolean add(long userId, long feedId, String content);
     boolean delete(long userId, long feedId, long commentId);
-    int queryCount(long feedId);
+    long queryCount(long feedId);
     List<FeedComment> query(long feedId, int start, int count);
 }
